% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charm.R
\name{charm}
\alias{charm}
\title{Encrypted HTML Documents}
\usage{
charm(input, password, hint, output = NULL, ...)
}
\arguments{
\item{input}{Path to an R Markdown or HTML file.}

\item{password}{Password to unlock file.}

\item{hint}{Public password hint.}

\item{output}{Name of the output file.}

\item{...}{Arguments passed on to \code{rmarkdown::render()}.}
}
\value{
\code{input}, invisibly.
}
\description{
Encrypt and password protect an HTML document.
}
