% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylize.R
\name{stylize}
\alias{stylize}
\title{Style password landing page}
\usage{
stylize(
  header_text = "PLEASE ENTER THE PASSWORD",
  placeholder_text = "",
  button_text = "SUBMIT",
  font_family = "-apple-system",
  font_size = "16px",
  title_color = "#2d3737",
  background_color = "#f2f2f2",
  box_color = "#ffffff",
  btn_font_color = "#ffffff",
  btn_bg_color = "#19A974",
  btn_hover_color = "#137752",
  modal_overlay = "rgba(0, 0, 0, 0.6)",
  modal_box_color = "#ffffff",
  modal_title_color = "#137752",
  modal_font_color = "rgba(0, 0, 0, 0.8)"
)
}
\arguments{
\item{header_text}{Header text.}

\item{placeholder_text}{Password input placeholder text.}

\item{button_text}{Button text.}

\item{font_family}{Font family.}

\item{font_size}{Font size.}

\item{title_color}{Title font color.}

\item{background_color}{Background color.}

\item{box_color}{Box color.}

\item{btn_font_color}{Button font color.}

\item{btn_bg_color}{Button background color.}

\item{btn_hover_color}{Button hover color.}

\item{modal_overlay}{Modal overlay color.}

\item{modal_box_color}{Modal box color.}

\item{modal_title_color}{Modal title font color.}

\item{modal_font_color}{Modal text font color.}
}
\value{
A list with class \code{fidelius_styling}.
}
\description{
Use this function to style the appearance of the landing
page for the password-protected HTML file.
}
\examples{

stylize(font_family = "Times")

}
