% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charm.R
\name{charm}
\alias{charm}
\title{Encrypted HTML Documents}
\usage{
charm(
  input,
  password,
  hint,
  output = NULL,
  style = stylize(),
  bundle = FALSE,
  ...
)
}
\arguments{
\item{input}{Path to an R Markdown or HTML file.}

\item{password}{Password to unlock file.}

\item{hint}{Public password hint.}

\item{output}{Name of the output file.}

\item{style}{Object returned from \code{stylize()}.}

\item{bundle}{Logical. Should all of the decryption machinery and
dependencies be bundled into the HTML document? Default is \code{FALSE}.}

\item{...}{Arguments passed on to \code{rmarkdown::render()}.}
}
\value{
\code{input}, invisibly.
}
\description{
Encrypt and password protect an HTML document.
}
\note{
Using \code{bundle = TRUE} only applies to bundling the
decryption machinery and dependencies for the document generated
by \code{charm()}. It is still the users responsibility to decide on whether
\code{input} is self-contained by passing \code{self_contained = TRUE} to
\code{rmarkdown::render()} using \code{...}, or by specifying
\code{self_contained = TRUE} in the YAML header of \code{input}.
}
