% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelius-package.R
\docType{package}
\name{fidelius-package}
\alias{fidelius}
\alias{fidelius-package}
\title{fidelius: Password Protected HTML Documents}
\description{
This package provides a simple interface to create
  secure, encrypted, password-protected static HTML documents, and creates
  self-contained HTML documents that contain the encrypted content
  and machinery for secure in-browser decryption.
}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
