% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidelius-package.R
\docType{package}
\name{fidelius-package}
\alias{fidelius}
\alias{fidelius-package}
\title{fidelius: Browser-Side Password-Protected HTML Documents}
\description{
Create secure, encrypted, and password-protected static HTML documents that include the machinery for secure in-browser authentication and decryption.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattwarkentin.github.io/fidelius/}
  \item \url{https://github.com/mattwarkentin/fidelius}
  \item Report bugs at \url{https://github.com/mattwarkentin/fidelius/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID}) [copyright holder]

}
\keyword{internal}
